% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{check_sample_names}
\alias{check_sample_names}
\title{Check sample names}
\usage{
check_sample_names(count_matrix, metadata, metadata_var)
}
\arguments{
\item{count_matrix}{preprocessed matrix of counts}

\item{metadata}{cleaned metadata for RNAseq data}

\item{metadata_var}{column of sample identifier that user expects to match with
count_matrix}
}
\value{
a `tbl` of the sorted metadata (if necessary), otherwise the supplied
metadata is returned with console message output of the quality control check
}
\description{
`check_sample_names()` is a simple quality control step that verifies whether
the column names in the count matrix match with a user-defined metadata column.
In order for a match to occur, the value levels of the column names and those of the
user-defined metadata column need to be identical
(e.g. `setdiff(colnames(my_count_matrix), metadata$my_column)` should be 0),
and the order in which the values appear need to be identical
(e.g. `identical(colnames(my_count_matrix), metadata$my_column)` should be TRUE).
}
\examples{

check_sample_names(counts, meta, FileName)

}
