% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{create_model_matrix}
\alias{create_model_matrix}
\title{Create a design matrix}
\usage{
create_model_matrix(metadata, vars)
}
\arguments{
\item{metadata}{cleaned metadata for RNAseq data}

\item{vars}{a character vector of variables to include in the model}
}
\value{
a `tbl` of the design matrix
}
\description{
`create_design_matrix()` creates a model matrix for your DE analysis.
See \code{\link[stats]{model.matrix}} for further details on what a design
(or model) matrix is.
}
\details{
The order in which you specify your variables will affect which
variable is dummy coded as the reference variable.
}
\examples{

make_design_matrix(data, "CellType")

make_design_matrix(data, c("CellType", "Status"))

# Results in a different design matrix
make_design_matrix(data, c("Status", "CellType"))
}
