% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter lowly expressed genes}
\usage{
filter_genes(count_matrix, filter_method)
}
\arguments{
\item{count_matrix}{preprocessed matrix of counts}

\item{filter_method}{Either `edgeR`, `samplenr`, or `cpm`}
}
\value{
a `tbl` of the design matrix
}
\description{
`filter_genes()` is a wrapper function for several filtering methods.
}
\details{
I encourage users to exercise caution before using this filter function.
Oftentimes, the filtering should be specific to the sequencing experiment.
The `edgeR` option is a wrapper for `edgeR::filterByExpr()`
}
