% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{make_contrasts}
\alias{make_contrasts}
\title{Make contrasts}
\usage{
make_contrasts(.fit, design_matrix, ...)
}
\arguments{
\item{.fit}{an MArrayLM object or list object. Must contain components
`coefficients` and `stdev.unscaled`}

\item{design_matrix}{a design matrix with rows corresponding to samples
and columns to coefficients to be estimated}

\item{...}{the names of the variables which you like to compare}
}
\value{
a `list` object of the same class as .fit
}
\description{
`make_contrasts()` computes estimated coefficients and standard errors
for a given set of contrasts
}
\details{
Please refer to \code{\link[limma]{contrasts.fit}} for more information.
}
\examples{
counts <- readr::read_delim("data/GSE60450_Lactation-GenewiseCounts.txt", delim = "\t")
meta <- readr::read_delim("data/SampleInfo_Corrected.txt", delim = "\t") \%>\%
  mutate(FileName = stringr::str_replace(FileName, "\\\\.", "-"))

id <- as.character(counts$EntrezGeneID)
check_sample_names(counts, c(1,2), meta, FileName) \%>\%
  purrr::pluck("mod_count") \%>\%
  filter_genes(., id, "edgeR") \%>\%
  make_voom(., my_design) \%>\%
  model_limma() \%>\%
  make_contrasts(Statuspregnant, Statusvirgin)
}
