% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Create a design matrix}
\usage{
make_design_matrix(metadata, vars)
}
\arguments{
\item{metadata}{cleaned metadata for RNAseq data}

\item{vars}{a character vector of variables to include in the model}
}
\value{
a `tbl` of the design matrix
}
\description{
`make_design_matrix()` creates a model matrix for your DE analysis.
See \code{\link[stats]{model.matrix}} for further details on what a design
(or model) matrix is.
}
\details{
The order in which you specify your variables will affect which
variable is dummy coded as the reference variable.
}
\examples{

make_design_matrix(metadata, "CellType")

make_design_matrix(metadata, c("CellType", "Status"))

# Results in a different design matrix
make_design_matrix(metadata, c("Status", "CellType"))

# In a pipeline
counts <- readr::read_delim("data/GSE60450_Lactation-GenewiseCounts.txt", delim = "\t")
meta <- readr::read_delim("data/SampleInfo_Corrected.txt", delim = "\t") \%>\%
  mutate(FileName = stringr::str_replace(FileName, "\\\\.", "-"))

my_design <- check_sample_names(counts, EntrezGeneID, c(1,2), meta, FileName) \%>\%
  purrr::pluck("meta") \%>\%
  make_design_matrix(., c("CellType"))
}
