% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{make_voom}
\alias{make_voom}
\title{Convert count data to voom}
\usage{
make_voom(.dge, design_matrix, .f = limma::voom, ...)
}
\arguments{
\item{.dge}{a `list` or a `DGElist` object}

\item{design_matrix}{a design matrix with rows corresponding to samples
and columns to coefficients to be estimated}

\item{.f}{limma::voom}

\item{...}{additional arguments passed to `.f`}
}
\value{
a `list` with the following components:
\item{E}{a numeric matrix of normalized expression values on the log2 scale}
\item{weights}{numeric matrix of inverse variance weights}
\item{design}{design matrix}
\item{lib.size}{numeric vector of total normalized library sizes}
\item{genes}{dataframe of gene annotation extracted from `counts`}
}
\description{
`make_voom()` is a wrapper function for `voom()`
}
\details{
Please refer to \code{\link[limma]{voom}} for more information.
}
\examples{
counts <- readr::read_delim("data/GSE60450_Lactation-GenewiseCounts.txt", delim = "\t")
meta <- readr::read_delim("data/SampleInfo_Corrected.txt", delim = "\t") \%>\%
  mutate(FileName = stringr::str_replace(FileName, "\\\\.", "-"))

id <- as.character(counts$EntrezGeneID)
check_sample_names(counts, c(1,2), meta, FileName) \%>\%
  purrr::pluck("mod_count") \%>\%
  filter_genes(., id, "edgeR") \%>\%
  make_voom(., my_design)
}
