% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{model_bayes}
\alias{model_bayes}
\title{Model differential expression}
\usage{
model_bayes(.fit, .f = limma::eBayes, ...)
}
\arguments{
\item{.fit}{an MArrayLM object or list object. Must contain components
`coefficients` and `stdev.unscaled`}

\item{.f}{limma::eBayes}

\item{...}{additional arguments to .f}
}
\value{
a `list` (`MArrayLM`) object containing the results of the fit
}
\description{
`model_bayes()` performs an empirical Bayes fit
}
\details{
Please refer to \code{\link[limma]{lmFit}} for more information.
}
\examples{

}
