% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{model_limma}
\alias{model_limma}
\title{Model limma}
\usage{
model_limma(.data, .f = limma::lmFit, ...)
}
\arguments{
\item{.f}{limma::lmFit}

\item{...}{additional arguments to .f}

\item{voom}{a voom object}
}
\value{

}
\description{
`model_limma()` is a wrapper function for `lmFit`. Fits a linear model for
each gene.
}
\details{
Please refer to \code{\link[limma]{lmFit}} for more information.
}
