% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{pca_plot}
\alias{pca_plot}
\title{Quickly plot a PCA plot}
\usage{
pca_plot(voom_data, metadata, ...)
}
\arguments{
\item{voom_data, }{voom object}

\item{metadata}{cleaned metadata for RNAseq data}

\item{...}{arguments to aes}
}
\value{
a scatterplot of the first two PC (ggplot object)
}
\description{
`pca_plot()` plots the first two PC in a scatterplot.
}
\details{
Like any other ggplot object, you can customize the theme of the plot.
Note that this function generates a fairly generic PCA plot, and thus is intended
for quick exploratory purposes (much like the intention behind `qplot()`).
}
\examples{
# As a EDA step within a pipeline of functions
check_sample_names(counts, c(1,2), meta, FileName) \%>\%
  purrr::pluck("mod_count") \%>\%
  filter_genes(., id, "edgeR") \%>\%
  make_voom(., my_design) \%T>\%
  {print(pca_plot(., check_sample_names(counts, c(1,2), meta, FileName) \%>\%
    purrr::pluck("meta")))} \%>\%
  model_limma() \%>\%
  make_contrasts(design_matrix = my_design, Statuspregnant, Statusvirgin) \%>\%
  model_bayes()
prcomp_obj, result from prcomp()
index, which PCA you want
}
