% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{run_sample_limma_de}
\alias{run_sample_limma_de}
\title{A function that runs a sample limma-voom DE workflow}
\usage{
run_sample_limma_de(
  raw_counts,
  metadata,
  .metadata_var,
  .design_vars,
  contrastcond1,
  contrastcond2,
  .cols_to_remove = NULL,
  .filter_method = "edgeR"
)
}
\arguments{
\item{raw_counts}{cleaned dataframe of counts, rows should be gene IDs,
columns should be samples, cells should only contain counts}

\item{metadata}{cleaned metadata for RNAseq data}

\item{.metadata_var}{column of sample identifier that user expects to match with
count_matrix}

\item{contrastcond1}{control var}

\item{contrastcond2}{experimentl var}

\item{.filter_method}{Please refer to \code{\link[tidyde]{filter_genes}} for more information.}

\item{cols_to_remove}{vector of column numbers that do not correspond to a sample,
necessary to identify for downstream functions}

\item{.design_var}{a character vector of variables to include in the model
these should be case-specific and match the variable names in the metadata}
}
\value{
a `tbl`
}
\description{
This wrapper function uses all the functions within this function and strings them
together in a sample workflow. Currently only `limma` is supported.
}
\examples{

cond_df <- tibble(cond1 = c("a1", "a2", "a3"), cond2 = c("b1", "b2", "b3"))

map2(cond_df$cond1, cond_df$cond2, ~run_sample_limma_de(
my_counts,
my_metadata,
my_variable,
my_design_variable,
.x,
.y))
}
