% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.marray.lm}
\alias{tidy.marray.lm}
\title{Tidying methods for MArrayLM objects}
\usage{
tidy.marray.lm(x, conf.int = TRUE, exponentiate = FALSE, ...)
}
\arguments{
\item{x}{a differential expression fit object. Currently supports `MArrayLM`
from the `limma` package.}

\item{conf.int}{logical. Include confidence intervals?}

\item{exponentiate}{logical. Should the estimates and (if `conf.int` = `TRUE`)
confidence intervals be exponentiated?}

\item{...}{additional arguments}
}
\value{
a `tbl`
}
\description{
These methods tidy the results of differential expression analysis objects.
Currently only `limma` is supported.
}
\examples{
counts <- readr::read_delim("data/GSE60450_Lactation-GenewiseCounts.txt", delim = "\t")
meta <- readr::read_delim("data/SampleInfo_Corrected.txt", delim = "\t") \%>\%
  mutate(FileName = stringr::str_replace(FileName, "\\\\.", "-"))

id <- as.character(counts$EntrezGeneID)
out <- check_sample_names(counts, c(1,2), meta, FileName) \%>\%
  purrr::pluck("mod_count") \%>\%
  filter_genes(., id, "edgeR") \%>\%
  make_voom(., my_design) \%>\%
  model_limma() \%>\%
  make_contrasts(Statuspregnant, Statusvirgin) \%>\%
  model_bayes()

out \%>\% tidy.marray.lm()

}
