% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{check_sample_names}
\alias{check_sample_names}
\title{Check sample names}
\usage{
check_sample_names(count_df, id_column, cols_to_remove, metadata, metadata_var)
}
\arguments{
\item{count_df}{cleaned dataframe of counts, rows should be gene IDs,
columns should be samples, cells should only contain counts}

\item{id_column}{identify the column name that contains the gene IDs}

\item{cols_to_remove}{vector of column numbers that do not correspond to a sample,
necessary to identify for downstream functions}

\item{metadata}{cleaned metadata for RNAseq data}

\item{metadata_var}{column of sample identifier that user expects to match with
count_matrix}
}
\value{
a `list` of `old_count`, `mod_count`, `id`, and `metadata`
where `old_count` is the original count dataframe supplied,
`mod_count` is the pure count dataframe (no other columns),
`id` is a vector of ID names,
and `metadata` is the sorted metadata (if necessary), otherwise the supplied
metadata is returned with console message output of the quality control check
}
\description{
`check_sample_names()` is a simple quality control step that verifies whether
the column names in the count matrix match with a user-defined metadata column.
In order for a match to occur, the value levels of the column names and those of the
user-defined metadata column need to be identical
(e.g. `setdiff(colnames(my_count_matrix), metadata$my_column)` should be 0),
and the order in which the values appear need to be identical
(e.g. `identical(colnames(my_count_matrix), metadata$my_column)` should be TRUE).
}
\examples{

check_sample_names(counts, EntrezGeneID, c(1,2), meta, FileName)

}
