% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{model_bayes}
\alias{model_bayes}
\title{Model differential expression}
\usage{
model_bayes(.fit, .f = limma::eBayes, ...)
}
\arguments{
\item{.fit}{an MArrayLM object produced by `model_limma()` or
`make_contrasts()`}

\item{.f}{limma::eBayes}

\item{...}{additional arguments to .f}
}
\value{
a `list` (`MArrayLM`) object containing the results of the fit
}
\description{
`model_bayes()` performs an empirical Bayes fit
}
\details{
Please refer to \code{\link[limma]{eBayes}} for more information.
}
\examples{
counts <- readr::read_delim("data/GSE60450_Lactation-GenewiseCounts.txt", delim = "\t")
meta <- readr::read_delim("data/SampleInfo_Corrected.txt", delim = "\t") \%>\%
  mutate(FileName = stringr::str_replace(FileName, "\\\\.", "-"))

id <- as.character(counts$EntrezGeneID)
check_sample_names(counts, c(1,2), meta, FileName) \%>\%
  purrr::pluck("mod_count") \%>\%
  filter_genes(., id, "edgeR") \%>\%
  make_voom(., my_design) \%>\%
  model_limma() \%>\%
  make_contrasts(Statuspregnant, Statusvirgin) \%>\%
  model_bayes()
}
